+++ apps/dashboard/layouts/Page.tsx
@@ -1,0 +1,6 @@
+import {ReactNode} from 'react';
+import * as s from './Page.css';
+
+export default function PageLayout({children}: {children: ReactNode}) {
+  return <section className={s.Page}>{children}</section>;
+}
